+++ lib/hcloud/entities/datacenter_server_type.rb
@@ -2,8 +2,19 @@
 
 module HCloud
   class DatacenterServerType < Entity
-    attribute :available, array: true, default: -> { [] }
-    attribute :available_for_migration, array: true, default: -> { [] }
-    attribute :supported, array: true, default: -> { [] }
+    attribute :available,
+              array: true,
+              default: -> { [] },
+              deprecated: true
+
+    attribute :available_for_migration,
+              array: true,
+              default: -> { [] },
+              deprecated: true
+
+    attribute :supported,
+              array: true,
+              default: -> { [] },
+              deprecated: true
   end
 end